make the main file shorter
