+++ src/COBREXA.jl
@@ -34,55 +34,14 @@ include_dependency(joinpath(_PKG_ROOT_DIR, "Project.toml"))
 const version =
     VersionNumber(Pkg.TOML.parsefile(joinpath(_PKG_ROOT_DIR, "Project.toml"))["version"])
 
-module ModuleTools
-macro inc(path...)
-    esc(:(include(joinpath(@__DIR__, $(joinpath(String.(path)...) * ".jl")))))
-end
+# bootstrap the module machinery
+include("modules.jl")
+using .ModuleTools
 
-macro inc_dir(path...)
-    dir = joinpath(@__DIR__, String.(path)...)
-    files = filter(endswith(".jl"), readdir(dir; join = true))
-    esc(Expr(:block, (:(include($f)) for f in files)...))
-end
-
-macro dse()
-    :(using DocStringExtensions)
-end
-
-macro inject(mod::Symbol, code)
-    esc(:(Base.eval($mod, $(Expr(:quote, code)))))
-end
-
-# export everything from the local namespace that seems exportable
-# (inspired by JuMP.jl, thanks!)
-macro export_locals()
-    quote
-        for sym in names(@__MODULE__; all = true, imported = true)
-            sym in [Symbol(@__MODULE__), :eval, :include] && continue
-            startswith(string(sym), ['_', '#']) && continue
-            sym == :Internal && continue
-            @eval export $(Expr(:$, :sym))
-        end
-    end
-end
-
-@export_locals
-end
-
-# load various internal helpers
-module Internal
-using ..ModuleTools
-@dse
-
-@inc macros
-@inc_dir misc ontology
-@inc_dir misc
-
-@export_locals
-end
+# load various internal helpers first
+@inc internal
 
 # start loading individual user-facing modules
-using .ModuleTools
 
 @inc log
 @inc types
