+++ src/internal.jl
@@ -1,0 +1,10 @@
+module Internal
+using ..ModuleTools
+@dse
+
+@inc macros
+@inc_dir misc ontology
+@inc_dir misc
+
+@export_locals
+end
